HTML/CSS HOMEWORK WEEK2
there is a margin right on the body when its viewed in smaller screens. its not very noticable until you reach 200px screen. i put width: fit-content; on the body which fixed the issue of the content overflowing when it got to very smaller screens. i tried max-width and even seached on stackoverflow but couldnt find a fix. i tried inspecting other sites in responsive mode but couldnt find the same problem. 

TLDR- there is a right margin on the body and i didnt declare it in the stylesheet.